refactor: extract status-bar.js, use direct import instead of argument passing
- status-bar.js: StatusBar class, getStatusBar(), wrap(),
  addLoadingIndicator() — app infrastructure used by all modules
- snapshotPoloidal() and generateSummary() now import getStatusBar
  directly instead of receiving it as a parameter
- openPanel() uses getStatusBar() import instead of DOM query
- index.js: 849 → 785 lines (64 removed)
