+++ CLAUDE.md
@@ -54,7 +54,8 @@ This is an Express.js app that visualizes GTC (Gyrokinetic Toroidal Code) simula
 
 ### Client-side (`public/`)
 
-- **`javascripts/index.js`** — Main module: `StatusBar` class, figure management, tab switching, routing to plot-specific handlers. Imports `state.js`, `util.js`, `api.js`, `units.js`, and `figure-range-controls.js`.
+- **`javascripts/index.js`** — Main module: figure lifecycle (openPanel, getDataThenPlot), bootstrap wiring, tab switching, breadcrumb navigation, download form. Imports `state.js`, `util.js`, `api.js`, `units.js`, `figure-range-controls.js`, and `status-bar.js`.
+- **`javascripts/status-bar.js`** — Application shell: `StatusBar` class (info/warn/err display), `getStatusBar()`, `wrap()` (async error handling), and `addLoadingIndicator()` (loading spinner wrapper). Used by virtually every other module; depends only on well-known DOM elements.
 - **`javascripts/state.js`** — Centralized application state object (replaces ad-hoc `window.GTCGlobal` properties). Documented with owning concern per property block (units, figure lifecycle, history mode, parameters, FFT). Also exposed as `window.GTCGlobal` for backward compatibility with modules not yet migrated to ES imports.
 - **`javascripts/util.js`** — Shared pure-utility functions with no app-state or DOM dependencies: `callEventTarget`, `propagateFetchError`, `nodeIs`, `postForm`, `min_max`, `interleave`, `unInterleave`.
 - **`javascripts/api.js`** — Server communication layer: `requestPlotData()` fetches plot data/metadata from the Express backend; `downloadOutputFiles()` POSTs a download request and returns the blob + filename. All fetch URL construction and error propagation lives here.
